提成

+++ src/main/java/com/gofobao/framework/collection/biz/impl/PaymentBizImpl.java
@@ -283,7 +283,7 @@ public class PaymentBizImpl implements PaymentBiz {
                     collectionDetail.setCollectionAt(DateHelper.dateToString(borrowCollection.getCollectionAt(), DateHelper.DATE_FORMAT_YMD));
                     collectionDetail.setPrincipal(StringHelper.formatDouble(borrowCollection.getPrincipal(), 100, false));
                     long money = borrowCollection.getInterest();
-                    if (borrow.getType() == 0 || borrow.getType() == 4) {
+                    if (borrow.getType().intValue() == 0 || borrow.getType().intValue() == 4) {
                         money -= new Double(MoneyHelper.round(borrowCollection.getInterest() * 0.9, 0)).longValue();
                     }
                     collectionDetail.setEarnings(StringHelper.formatDouble(money, 100, false));
